Stop relying only on isHubCluster to enable the unsealJob
Eventually we would like to drop isHubCluster as it is confusing and its
meaning does not always make sense (think of a standalone mcg setup.
I.e. one without ACM at all).

So let's enable the unsealJob either when isHubCluster is set to true
(in order to not break compatibility) or when in the application we
have the "hashicorp-vault" chart.

Tested on a standalone target for mcg and everything worked correctly.